Removing useless component and updating buttons behaviour
