--- src/double-action-button/style/double-action-button.scss
@@ -1,6 +1,0 @@
-
-[data-focus="button-action"] {
-    i {
-        margin-right: 5px;
-    }
-}
