+++ src/index.js
@@ -17,7 +17,6 @@ import './chips';
 import './column';
 import './confirmation-popin';
 import './draggable-iframe';
-import './double-action-button';
 import './dropdown';
 import './empty';
 import './form';
